+++ demo/stores/app.js
@@ -12,6 +12,11 @@ let data = ImmutableHelper.parseJSON({
   collapsible_pod: {
     title: 'Hello World!'
   },
+  dialog: {
+    open: false,
+    title: 'Hello World!',
+    disableBackground: true
+  },
   flash: {
     as: "warning",
     open: true,
@@ -30,6 +35,15 @@ let data = ImmutableHelper.parseJSON({
   pill: {
     as: 'info',
     text: 'Pill'
+  },
+  portrait: {
+    size: 'lmed',
+    shape: 'standard',
+    email: 'Email'
+  },
+  pod: {
+    border: true,
+    padding: "medium"
   },
   rainbow: {
     data: [{
@@ -79,6 +93,9 @@ let data = ImmutableHelper.parseJSON({
   },
   tabs: {
     tabData: [{}, {}]
+  },
+  tile: {
+    content: "Using a borderless Pod can provide padding inside your tile."
   },
   toast: {
     as: "warning",